history_sync: Further dividing, thank god I did it easy
So this boring shit is not broken up in pieces, it should still be functional,
but we need some kind of functional tests I am afraid. There is too much
logic involved.

Also there are a few optimizations that I would like to do
because I have seen that there are quite a few places with the same lines
of "get or create" and I think I did it because it was more performant
that way, but need to revisit it.

I am also not sure if underlying implementation of all the saves should
change or not